CHANGE: Break up UnityEngine.Input-tests into multiple files.
