+++ Assets/Tests/InputSystem/CoreTests_API_Joystick.cs
@@ -1,0 +1,11 @@
+using NUnit.Framework;
+
+partial class CoreTests
+{
+    [Test]
+    [Category("API")]
+    public void API_CanListJoysticksAndGamepadsThroughGetJoystickNamesAPI()
+    {
+        Assert.Fail();
+    }
+}
